Quiz is Semi-Working !!!!
Yay TT

Still some issues though. Gonna need to impement channel layers of sorts

Maybe some logic needs to be re-assessed so that people don't reload the page by loading in (on connect)

Scoring is messed up in multiplayer. i think the scores should be added to a dictionary and adjusted when each person selects an options. And then then once all the scores are added they should all be processed from the dictionary and broadcast