Websocket and DjangoToolbar update
Updated the WebSocket URL to use dynamic protocol selection. This should hopefully alleviate the ws and wss issues we were having. Changes were made in quiz.js and in routing.py

I also installed Django Toolbar for debugging and optimising the site. I had to change some things in settings and requirements.txt inorder for it be activated during localserver run testing, but not on heroku server deployment